fix(wrapper,auth,middleware): address PR-A11 review Cx1/Cx2 findings
- F-02 correctness: http.status_code attribute now int64 on both the
  outer middleware.Tracing span and the inner kernel/wrapper span, so
  OTel collector pipelines that switch on attribute value type see a
  consistent representation across sibling spans. tracing_test.go
  assertion updated.

- F-03 DX: godoc on kernel/wrapper.HTTPHandler explains the two-span
  co-existence (outer chi-route-pattern span + inner contract span)
  that results from composing with runtime/http/middleware.Tracing,
  so operators reading the code understand dashboard double-counting
  before they hit it.

- F-04 DX: runtime/auth.Mount now carries an inline explanation of
  why RequirePolicy wraps OUTERMOST — policy denials must short-circuit
  before the contract span is started, otherwise every pre-auth reject
  floods observability backends with spans.

- F-05 DX: Route.Contract godoc teases apart the two Method/Path pairs
  — Route.Method/Path drive mux registration, Contract.Method/Path
  drive span attributes — so future developers who look at the struct
  don't wonder why the duplication exists.

- F-08 testing: TestMount_PathNormalised is now table-driven with four
  cases including /login/ → /login to document the trailing-slash
  stripping behavior that path.Clean inherits.

- F-10 correctness / cleanup: statusRecorder.Write no longer re-assigns
  status = http.StatusOK (it was already seeded at construction); the
  updated comment explains the invariant so readers don't think Write
  is authoritative for the captured status.

- F-07 ack: cells/accesscore/cell_routes.go carries a TODO(FMT-17)
  comment pointing at the governance rule that will cross-check this
  literal against contracts/http/auth/login/v1/contract.yaml once
  PR-A11-V lands.

F-01 confirmed false-positive: kernel/wrapper/correlation.go DOES
import pkg/ctxkeys (for TraceID/SpanID bridges), so doc.go's
dependency claim is accurate.

F-09 already covered: TestWrapConsumer_InvalidDisposition_MarksError
WithoutModifyingResult in handler_chi_test.go documents the "emit
StatusError span but do not modify HandleResult" semantics.

Refs: PR-A11 / LATER-K1 KERNEL/WRAPPER

+++ runtime/http/middleware/tracing.go
@@ -100,9 +100,13 @@ func Tracing(tracer tracing.Tracer, opts ...TracingOption) func(http.Handler) ht
 			tracing.SpanSetName(span, r.Method+" "+route)
 
 			status := state.Status()
+			// Emit http.status_code as int64 for cross-span type consistency —
+			// kernel/wrapper.HTTPHandler uses int64 on the inner contract span,
+			// and OTel collector pipelines that switch on attribute type must
+			// see the same type across sibling spans in one trace. (F-02)
 			span.SetAttributes(
 				tracing.Attr{Key: "http.route", Value: route},
-				tracing.Attr{Key: "http.status_code", Value: status},
+				tracing.Attr{Key: "http.status_code", Value: int64(status)},
 			)
 
 			// 5xx → error span; 4xx and below → unset (otelhttp convention).
